requirements.txt
aiofiles
aiohttp
aiomqtt
fastapi
pycryptodome
pytest
pytest-asyncio
uvicorn
websockets
python-dotenv